raw match object moved under GameHandler package
