+++ src/BSC/DataExtractor/fromTS.py
@@ -2,6 +2,8 @@ import re
 import requests
 from bs4 import BeautifulSoup
 
+from BSC.GameHandler.rawmatch import RawMatch
+
 HEADERS = {
     "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
     "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,*/*;q=0.8",
@@ -134,21 +136,3 @@ class WebScraper():
 
                 self.matches_list.append(match_dict)
                 self.rawMatchesObjects_list.append(new_raw_match)
-
-
-class RawMatch():
-    def __init__(self, category, league, team_one, team_one_status, team_one_scores, team_two, team_two_status, team_two_scores):
-        self.category = category
-        self.league = league
-        self.team_one = team_one
-        self.team_one_status = team_one_status
-        self.team_one_scores = team_one_scores
-        self.team_two = team_two
-        self.team_two_status = team_two_status
-        self.team_two_scores = team_two_scores
-
-    def GetMatchString(self):
-        result_dict = {}
-        result_dict[self.team_one] = self.team_one_scores
-        result_dict[self.team_two] = self.team_two_scores
-        return result_dict
